+++ src/main/java/com/movietix/xiazihao/entity/param/MovieQueryParam.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -16,8 +15,8 @@ public class MovieQueryParam {
     private Integer minDuration;   // 最小时长(分钟)
     private Integer maxDuration;   // 最大时长(分钟)
     private String genre;          // 电影类型
-    private BigDecimal minRating;  // 最低评分
-    private BigDecimal maxRating;  // 最高评分
+    private Double minRating;  // 最低评分
+    private Double maxRating;  // 最高评分
     private Integer status;        // 状态(0-下架,1-上映)
     private Integer page = 1;      // 页码(默认1)
     private Integer pageSize = 10; // 每页条数(默认10)
