+++ src/main/java/com/movietix/xiazihao/entity/pojo/OrderSeat.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -15,5 +13,5 @@ public class OrderSeat {
     private Integer seatRow; // 座位行
     private Integer seatCol; // 座位列
     private String seatNo; // 座位号(如A01)
-    private BigDecimal price; // 座位价格
+    private Double price; // 座位价格
 }
