+++ src/main/java/com/movietix/xiazihao/entity/pojo/Payment.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -14,7 +13,7 @@ import java.util.List;
 public class Payment {
     private Integer id; // 主键ID
     private Integer orderId; // 订单ID
-    private BigDecimal amount; // 支付金额
+    private Double amount; // 支付金额
     private String paymentMethod; // 支付方式
     private String transactionId; // 交易号
     private Integer status = 0; // 状态:0-未支付 1-支付成功 2-支付失败
@@ -23,6 +22,6 @@ public class Payment {
     private LocalDateTime updatedAt; // 更新时间
 
     private String contactPhone; // 联系电话
-    private BigDecimal totalAmount; // 总金额
+    private Double totalAmount; // 总金额
     private List<OrderSeat> seats; // 座位号列表
 }
