+++ src/main/java/com/movietix/xiazihao/entity/pojo/Refund.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -18,7 +17,7 @@ public class Refund {
     private Integer status = 0; // 状态:0-待处理 1-已同意 2-已拒绝
     private Integer adminId; // 处理管理员ID
     private LocalDateTime processedAt; // 处理时间
-    private BigDecimal refundAmount; // 退款金额
+    private Double refundAmount; // 退款金额
     private LocalDateTime createdAt; // 创建时间
     private LocalDateTime updatedAt; // 更新时间
