+++ src/main/java/com/movietix/xiazihao/entity/pojo/Screening.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -16,7 +15,7 @@ public class Screening {
     private Integer hallId; // 放映厅ID
     private LocalDateTime startTime; // 开始时间
     private LocalDateTime endTime; // 结束时间
-    private BigDecimal price; // 票价
+    private Double price; // 票价
     private Integer remainingSeats; // 剩余座位数
     private Integer status = 1; // 状态:0-已取消 1-正常
     private LocalDateTime createdAt; // 创建时间
