+++ src/main/java/com/movietix/xiazihao/entity/result/WorkSeatResult.java
@@ -5,7 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -14,6 +13,6 @@ import java.util.List;
 public class WorkSeatResult {
     private Integer rows; // 总行数
     private Integer cols; // 总列数
-    private BigDecimal price; //单价
+    private Double price; //单价
     List<OrderSeat> seats;
 }
